refactor: :truck: extract shipping transformer
